Endret dokumentadapter til å hente liste over dokumenter i stedet for liste av strenger
Lagt på tester
Endret packetKeys fra dokumentTitler til dokumenter
