+++ src/test/kotlin/no/nav/dagpenger/journalføring/arena/JournalFøringArenaTopologyTest.kt
@@ -67,7 +67,7 @@ class JournalFøringArenaTopologyTest {
             )
             putValue("naturligIdent", "12345678")
             putValue("journalpostId", "666")
-            putValue("dokumentTitler", listOf(JournalføringArena.Dokument("tittel 1")))
+            putValue("dokumenter", listOf(Dokument("Søknad 1")))
             putValue("datoRegistrert", "2019")
         }
